move `window.showOpenFilePicker` to tentative
