+++ webdriver/tests/bidi/input/file_dialog_opened/__init__.py
@@ -1,7 +1,8 @@
 from tests.bidi import recursive_compare, any_bool, any_dict
 
 
-def assert_file_dialog_opened_event(event, context, multiple=any_bool, element=any_dict):
+def assert_file_dialog_opened_event(event, context, multiple=any_bool,
+        element=any_dict):
     recursive_compare({
         'context': context,
         'element': element,
